feat(core): protection login page using internal styles
