+++ packages/core/src/pages/fs-auth-login/fs-auth-login.module.scss
@@ -1,0 +1,41 @@
+@layer components {
+  .fsAuthLogin {
+    @import "@faststore/ui/src/components/atoms/Button/styles.scss";
+    @import "@faststore/ui/src/components/atoms/Input/styles.scss";
+    @import "@faststore/ui/src/components/atoms/Loader/styles.scss";
+    @import "@faststore/ui/src/components/molecules/InputField/styles.scss";
+  }
+}
+
+.page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+  padding: var(--fs-spacing-3);
+}
+
+.title {
+  margin: 0 0 var(--fs-spacing-6) 0;
+  font-size: var(--fs-text-size-title-section);
+  font-weight: var(--fs-text-weight-bold);
+}
+
+.subtitle {
+  margin: 0 0 var(--fs-spacing-3) 0;
+  font-size: var(--fs-text-size-body);
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: var(--fs-spacing-2);
+  align-items: stretch;
+  width: 100%;
+  max-width: 20rem;
+
+  button {
+    align-self: center;
+  }
+}
